изменения по замечаниям ревью

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -14,12 +14,12 @@ public class Film {
     public final static int DESCRIPTION_MAX_LENGTH = 200;
     public final static LocalDate MIN_RELEASE_DATE = LocalDate.of(1895, 12, 28);
 
-    private int id;
+    private long id;
     @NotBlank
     private String name;
     private String description;
     private LocalDate releaseDate;
     @Positive
     private int duration;
-    private HashSet<Integer> likes;
+    private HashSet<Long> likes;
 }
